Add transaction deletion confirmation
